+++ trunk/SMS/SMS.Services/IInvoiceDetailService.cs
@@ -5,7 +5,7 @@ namespace SMS.Services
 {
     public interface IInvoiceDetailService : IBaseService<InvoiceDetailDto, long>
     {
-        InvoiceDetailDto AddProductToInvoiceTable(long invoiceTableID, long productID, int quantity);
+        InvoiceDetailDto AddProductToInvoiceTable(long invoiceTableID, long productID, decimal quantity);
         InvoiceDetailDto UpdateProductToInvoiceTable(long invoiceDetailID, string columnName, string value);
     }
 }
